+++ toys-core/test/definition/test_flag.rb
@@ -265,10 +265,18 @@ describe Toys::Definition::FlagSyntax do
 end
 
 describe Toys::Definition::Flag do
-  it "defaults to a boolean switch" do
+  it "defaults to a boolean switch with a long name" do
     flag = Toys::Definition::Flag.new(:abc, [], [], true, nil, nil, nil, nil, nil, nil)
     assert_equal(1, flag.flag_syntax.size)
     assert_equal("--abc", flag.flag_syntax.first.canonical_str)
+    assert_equal(:boolean, flag.flag_syntax.first.flag_type)
+    assert_equal(:boolean, flag.flag_type)
+  end
+
+  it "defaults to a boolean switch with a short name" do
+    flag = Toys::Definition::Flag.new(:a, [], [], true, nil, nil, nil, nil, nil, nil)
+    assert_equal(1, flag.flag_syntax.size)
+    assert_equal("-a", flag.flag_syntax.first.canonical_str)
     assert_equal(:boolean, flag.flag_syntax.first.flag_type)
     assert_equal(:boolean, flag.flag_type)
   end
@@ -304,6 +312,22 @@ describe Toys::Definition::Flag do
     assert_equal("VALUE", flag.value_label)
     assert_nil(flag.default)
     assert_equal(acceptor, flag.acceptor)
+  end
+
+  it "defaults to a value switch with a short name" do
+    flag = Toys::Definition::Flag.new(:a, [], [], true, nil, nil, "hello", nil, nil, nil)
+    assert_equal(1, flag.flag_syntax.size)
+    assert_equal("-a VALUE", flag.flag_syntax.first.canonical_str)
+    assert_equal(:value, flag.flag_syntax.first.flag_type)
+    assert_equal(:required, flag.flag_syntax.first.value_type)
+    assert_equal(" ", flag.flag_syntax.first.value_delim)
+    assert_equal("VALUE", flag.flag_syntax.first.value_label)
+    assert_equal(:value, flag.flag_type)
+    assert_equal(:required, flag.value_type)
+    assert_equal(" ", flag.value_delim)
+    assert_equal("VALUE", flag.value_label)
+    assert_equal("hello", flag.default)
+    assert_nil(flag.acceptor)
   end
 
   it "chooses the first long flag's value label and delim as canonical" do